+++ LINBIT/PluginHelper.pm
@@ -48,25 +48,4 @@ sub get_images {
     return $res;
 }
 
-sub get_status {
-    my ( $storage_pools, $storage_pool, $node_name ) = @_;
-    #     overall      , filter      , node_filter
-
-    my ( $avail_kib, $total_kib );
-    foreach my $name ( keys %$storage_pools ) {
-        next if $storage_pool and 0 == ( scalar grep { $_ eq $name } @$storage_pool );
-
-        # skip if not on this node
-        next unless exists $storage_pools->{$name}->{$node_name};
-
-        # skip diskless pools. LINSTOR considers them having infinite space, so they have some MAX values.
-        next if $storage_pools->{$name}->{$node_name}->{conf_as_diskless};
-
-        $avail_kib += $storage_pools->{$name}->{$node_name}->{free_capacity_kib};
-        $total_kib += $storage_pools->{$name}->{$node_name}->{total_capacity_kib};
-    }
-
-    return ( $total_kib, $avail_kib );
-}
-
 1;