+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- switch to cluster wide storage view instead of node-local view (requires LINSTOR >= 1.21.1)
 
 ## [6.1.0] - 2022-10-11
 ### Added